Onboarding folder restructuring
